Tests/reorganize: move custom header tests to own file
